Do not use function name anywhere
